coordenadas/requirements.txt
kafka-python>=2.0.2
asyncio>=3.4.3
aiokafka>=0.7.2
threaded>=4.1.0